amazon linuxの場合おかしくなっていたので変更

+++ setup.sh
@@ -5,13 +5,20 @@ mkdir -p /etc/logmon
 pushd $(dirname $0)
 cp ./logmon.pl /etc/logmon/
 cp ./logmon.conf /etc/logmon/
-cp ./supervisor_logmon.conf /etc/supervisor/conf.d/logmon.conf
+if [ -e "/etc/ec2_version" ]; then
+  cp ./supervisor_logmon.conf /etc/supervisor/conf.d/logmon.conf
+fi
 popd
 
 chmod 700 /etc/logmon/logmon.pl
 chmod 600 /etc/logmon/logmon.conf
-chmod 600 /etc/supervisor/conf.d/logmon.conf
 
-service supervisor start
-supervisorctl reread
-supervisorctl start logmon
+if [ -e "/etc/ec2_version" ]; then
+  chmod 600 /etc/supervisor/conf.d/logmon.conf
+fi
+
+if [ -e "/etc/ec2_version" ]; then
+  service supervisor start
+  supervisorctl reread
+  supervisorctl start logmon
+fi
